+++ InputGlyphs/Assets/InputGlyphs/Scripts/Runtime/Attributes/ControlSchemeNameAttribute.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+namespace InputGlyphs.Attributes
+{
+    public class ControlSchemeNameAttribute : PropertyAttribute
+    {
+    }
+}